requirements.txt
aiogram>=3.0.0
pytesseract
opencv-python-headless
numpy
python-dotenv